Special-Offers/requirements.txt
ruamel.yaml
requests
lxml
html5lib
beautifulsoup4
selenium